day 18 part 2 optimizations
change stack to unordered_set. The stack could contain up to 6 repeats
of the same coordinate to check for filling. This also changes the dfs
flood fill into a random walk flood fill.

Time for both parts goes down from ~17ms to ~14ms, the maximum set of
yet to examine coordinates goes down from 14259 to 3865 (from 244 to 58
on the example.txt).